refactoring:move formateDate function to utils
